+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * This fixes any "SBAT self-check failed" errors to ensure Rescuezilla continues boot on UEFI Secure Boot enabled systems which require the latest "SBAT generation"
   * This also fixes the "revoked UEFI bootloader" message warning when creating a bootable USB stick using Rufus
 * Replaces Ubuntu 23.10 (Mantic) and Ubuntu 22.10 (Lunar) builds with build based on Ubuntu 24.10 (Oracular) for best support of new hardware 
+* Fixed querying drives with the Bionic 32-bit version, which broke since Rescuezilla v2.5 due to using the --merge feature introduced in util-linux v2.34 ([#509](https://github.com/rescuezilla/rescuezilla/issues/509), [#531](https://github.com/rescuezilla/rescuezilla/issues/531))
 * Skips GPG check on Bionic 32-bit release to fix build (temporarily) until better solution identified ([#538](https://github.com/rescuezilla/rescuezilla/issues/538))
 * Upgraded to latest partclone `v0.3.33` (released mid-July 2024) from partclone `v0.3.27` (released October 2023)
 * Upgrades memtest86+ v5.31 to memtest86+ v7.00 ([#540](https://github.com/rescuezilla/rescuezilla/issues/540))